trim exchanges using current, rather than parent id; move conversation load and store messages to conversation.rs
